fix: remove ability for editors to edit relationships
Relationships are pretty broken right now. The UI doesn't work because
of the switch to ReactSelect, the backend doesn't prevent stupid
relationships, and the code is bloated and has a lot of debt. So there
will be a rewrite in the near future using Redux and a schema change,
but for now, I don't want to put in a lot of work to patch the mess.