fix(search): reimplement getEntity utility
This was called getEntityAlias for some reason, while it fetches the entire entity rather than only the alias(es). Not sure why.

Reimplemented it to allow passing ORM fetch options (in this case to set require:false and transaction object).

It would be worth modifying the similarly named utility in the ORM to allow passing options, but I'm in a rush to fix a pressing issue.