chore(search): Index only the required fields
We have been storing a ton of information in the search index that we just won't ever use such as set ids and internal props from the ORM Model (_pivot…).
Instead, let's pass the ORM models along and create a new utility to strip the dcument to index down to what we actually need.