fix: Only create a new revision if there are actual changes
In case we want to create a new annotation, we initially have to create
it without an associated revision and can later update `latestRevision`.

Without this patch, the entity editors can not be used to cleanly
"approve" pending entities without creating an empty revision.

An alternative idea which has been considered was to rollback the
transaction in that case, but this has annoying side effects:
- The ID of the rolled back revision is still lost and will be unused.
- It seems impossible to return the main entity for indexing.