Add margin and controlId to FromGroup elements
with the removal of the .form-group class and styling, we need to use utility classes to add back the margin, according to the migration docs (https://react-bootstrap.github.io/docs/migrating/#formgroup).

Also adding controlId for each field for accessibility. It adds the id and for props to the label and input respectively.