fix: Improve error catching in entity loader
When any database error is thrown in the entity loader, the actual error message is hidden and replaced with a generic "not found" error as the code seems to assume any error should be a notfound.
Check the type of error and return an actual error message in all other cases so we can actually debug what is happening.